+++ code/src/cljs/sixsq/nuvla/ui/deployment_dialog/views_price.cljs
@@ -4,6 +4,7 @@
     [re-frame.core :refer [dispatch subscribe]]
     [sixsq.nuvla.ui.deployment-dialog.events :as events]
     [sixsq.nuvla.ui.deployment-dialog.subs :as subs]
+    [sixsq.nuvla.ui.deployment-dialog.utils :as utils]
     [sixsq.nuvla.ui.i18n.subs :as i18n-subs]
     [sixsq.nuvla.ui.utils.general :as general-utils]
     [sixsq.nuvla.ui.utils.semantic-ui :as ui]
@@ -12,10 +13,10 @@
 
 (defn summary-row
   []
-  (let [tr                 (subscribe [::i18n-subs/tr])
+  (let [tr               (subscribe [::i18n-subs/tr])
         price            (subscribe [::subs/price])
         price-completed? (subscribe [::subs/price-completed?])
-        on-click-fn        #(dispatch [::events/set-active-step :pricing])]
+        on-click-fn      #(dispatch [::events/set-active-step :pricing])]
     [ui/TableRow {:active   false
                   :on-click on-click-fn}
      [ui/TableCell {:collapsing true}
@@ -28,19 +29,35 @@
                            "€/" (str/capitalize (@tr [:month])))]]]))
 
 
-(defn content
+(defmethod utils/step-content :pricing
   []
   (let [tr               (subscribe [::i18n-subs/tr])
-        price            (subscribe [::subs/price])
         price-completed? (subscribe [::subs/price-completed?])
         coupon           (subscribe [::subs/coupon])
-        deployment       (subscribe [::subs/deployment])]
+        deployment       (subscribe [::subs/deployment])
+        start?           (subscribe [::subs/deployment-start?])
+        price            (subscribe [::subs/price])
+        new-price        (subscribe [::subs/new-price])
+        format-price     #(if (>= (:cent-amount-daily %) 100)
+                            (str (float (/ (:cent-amount-daily %) 100)) "€/" (@tr [:day]))
+                            (str (:cent-amount-daily %) "ct€/" (@tr [:day])))]
     [:<>
      [ui/Segment
-      [:p (str (@tr [:one-day-trial-deployment])) " " (@tr [:deployment-will-cost])
-       [:b (if (>= (:cent-amount-daily @price) 100)
-             (str (float (/ (:cent-amount-daily @price) 100)) "€/" (@tr [:day]))
-             (str (:cent-amount-daily @price) "ct€/" (@tr [:day])))]]
+      [:p
+       (str (when @start? (@tr [:trial-deployment]))
+            (cond-> (@tr [:deployment-will-cost])
+                    @new-price (str/capitalize)))
+       (when @new-price
+         (if (> (:cent-amount-daily @new-price)
+                (:cent-amount-daily @price))
+           [ui/Icon {:name "caret up" :color "red"}]
+           [ui/Icon {:name "caret down" :color "green"}]))
+
+       [:b (format-price (or @new-price @price))]]
+      (when @new-price
+        [:p [:i (str (@tr [:price-changed])
+                  (format-price @price) " " (@tr [:to]) " "
+                  (format-price @new-price))]])
       [ui/Checkbox {:label     (@tr [:accept-costs])
                     :checked   @price-completed?
                     :on-change (ui-callback/checked